Fix the highlight remaining issue.
