+++ src/app/arb_validation/arb_validation.repository.ts
@@ -1,7 +1,8 @@
 import * as vscode from "vscode";
 import { Arb } from "../arb/arb";
 import { BaseDisposable } from "../util/base/base_disposable";
-import { Editor, HighlightType } from "../util/editor";
+import { Editor } from "../util/editor";
+import { Highlight, HighlightType } from "../util/highlight";
 import {
   ArbValidation,
   ArbValidationData,
@@ -63,29 +64,33 @@ export class ArbValidationRepository extends BaseDisposable {
    */
   public async keyRequired(sourceArb: Arb, targetArb: Arb, key: string) {
     try {
-      // source
+      // clear remain decorations
+      Highlight.clear();
+
+      // open document
       const { editor: sourceEditor, document: sourceDocument } =
         await Editor.open(sourceArb.filePath, vscode.ViewColumn.One);
+      const { editor: targetEditor, document: targetDocument } =
+        await Editor.open(targetArb.filePath, vscode.ViewColumn.Two);
+
+      // search key
       const sourceKeyPosition = Editor.search(sourceEditor, key);
-      const {
-        range: sourceHighlightedRange,
-        decorationType: sourceDecorationType,
-      } = Editor.highlight(
+
+      // highlight
+      const sourceHighlightedRange = Highlight.add(
         sourceEditor,
         HighlightType.green,
         sourceKeyPosition!.line
       );
 
       // target
-      const { editor: targetEditor, document: targetDocument } =
-        await Editor.open(targetArb.filePath, vscode.ViewColumn.Two);
       targetEditor.revealRange(
         sourceHighlightedRange,
         vscode.TextEditorRevealType.InCenter
       );
 
       const removeHighlight = () => {
-        Editor.clearHighlight(sourceEditor, sourceDecorationType);
+        Highlight.clear();
         this.disposed();
       };
 
@@ -122,31 +127,26 @@ export class ArbValidationRepository extends BaseDisposable {
     sourceArbValidationData: ArbValidationData
   ) {
     try {
-      // source
+      // clear remain decorations
+      Highlight.clear();
+
+      // open document
       const { editor: sourceEditor, document: sourceDocument } =
         await Editor.open(sourceArb.filePath, vscode.ViewColumn.One);
-      const sourceKeyPosition = Editor.search(sourceEditor, key);
-      const {
-        range: sourceHighlightedRange,
-        decorationType: sourceDecorationType,
-      } = Editor.highlight(
-        sourceEditor,
-        HighlightType.green,
-        sourceKeyPosition!.line
-      );
-
-      // target
       const { editor: targetEditor, document: targetDocument } =
         await Editor.open(targetArb.filePath, vscode.ViewColumn.Two);
+
+      // search key
+      const sourceKeyPosition = Editor.search(sourceEditor, key);
       const targetKeyPosition = Editor.search(targetEditor, key);
       if (!targetKeyPosition) {
         return;
       }
-      const { decorationType: targetDecorationType } = Editor.highlight(
-        targetEditor,
-        HighlightType.red,
-        targetKeyPosition.line
-      );
+
+      // highlight
+      Highlight.add(sourceEditor, HighlightType.green, sourceKeyPosition!.line);
+
+      Highlight.add(targetEditor, HighlightType.red, targetKeyPosition.line);
 
       // select target value
       const targetValue = targetArb.data[key];
@@ -157,8 +157,7 @@ export class ArbValidationRepository extends BaseDisposable {
       );
 
       const removeHighlight = () => {
-        Editor.clearHighlight(sourceEditor, sourceDecorationType);
-        Editor.clearHighlight(targetEditor, targetDecorationType);
+        Highlight.clear();
         this.disposed();
       };
